requirements.txt
Flask>=2.3.3,<3.0.0
Flask-SQLAlchemy>=3.0.0    # Updated to newer version compatible with SQLAlchemy 2.x
SQLAlchemy>=2.0.0          # Updated to 2.x series which supports Python 3.13
pandas>=2.0.0              # Updated to version supporting Python 3.13
scipy>=1.11.0              # Updated to version supporting Python 3.13
scikit-learn>=1.3.0          # Added for linear regression etc.
statsmodels>=0.14.0          # Added for more advanced stats/seasonality (optional for now)
numpy>=1.25.0              # Updated to version supporting Python 3.13
gunicorn==20.1.0
python-dotenv==0.21.0
Werkzeug>=2.3.8,<3.0.0     # Compatible with Flask 2.3.x
Jinja2>=3.1.2
itsdangerous>=2.1.2
MarkupSafe>=2.1.1
requests>=2.31.0
Flask-Migrate>=4.0.5       # Works with newer SQLAlchemy
psycopg2-binary>=2.9.9
boto3==1.38.20
rapidfuzz>=3.6.1